Выносит типы предложений в отдельный файл и меняет название функции получения моковых данных
